+++ example/lib/custom_stepper_demo.dart
@@ -36,9 +36,7 @@ class _CustomStepperDemoWidgetState extends State<CustomStepperDemoWidget> {
           width: 200,
           height: 25,
           stepCount: 3,
-          padding: 5,
-          builder: (int index) {
-            const double widthOfStep = 200.0 / 3.0;
+          builder: (int index, double widthOfStep) {
             if (index == 1) {
               return ProgressStepWithArrow(
                 width: widthOfStep,
@@ -47,14 +45,11 @@ class _CustomStepperDemoWidgetState extends State<CustomStepperDemoWidget> {
                 progressColor: Colors.green,
                 borderWidth: 1,
                 wasCompleted: _customCounter >= index,
-                child: SizedBox(
-                  width: widthOfStep,
-                  child: Center(
-                    child: Text(
-                      index.toString(),
-                      style: const TextStyle(
-                        color: Colors.white,
-                      ),
+                child: Center(
+                  child: Text(
+                    index.toString(),
+                    style: const TextStyle(
+                      color: Colors.white,
                     ),
                   ),
                 ),
@@ -67,15 +62,12 @@ class _CustomStepperDemoWidgetState extends State<CustomStepperDemoWidget> {
               progressColor: Colors.green,
               borderWidth: 1,
               wasCompleted: _customCounter >= index,
-              child: SizedBox(
-                width: widthOfStep,
-                child: Center(
-                  child: Text(
-                    index.toString(),
-                    textAlign: TextAlign.center,
-                    style: const TextStyle(
-                      color: Colors.white,
-                    ),
+              child: Center(
+                child: Text(
+                  index.toString(),
+                  textAlign: TextAlign.center,
+                  style: const TextStyle(
+                    color: Colors.white,
                   ),
                 ),
               ),